backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
# Pin networkx so install doesn't hang on resolution (langgraph/langchain depend on it)
networkx>=3.0,<4
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
sqlite-vec>=0.1.0
pysqlite3>=0.5.0
python-dotenv>=1.0.0
typing_extensions>=4.0.0
# Optional: only used when LIGHTRAG_ENABLED=true (see lightrag_bridge.py)
lightrag-hku>=1.3.0,<1.4.0
numpy>=1.20.0
tavily-python>=0.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
pytest>=8.0.0
